Create new RateTimeInTransit class for new methods
